Move authentication-related endpoints into a separate module
