+++ backend/src/api/endpoints/user.py
@@ -1,7 +1,6 @@
 from typing import Dict, List, Optional
 
 from api.dependencies import get_current_user, get_db_session, skip_limit_params
-from api.util import credentials_exception
 from app.core.data.crud.annotation_document import crud_adoc
 from app.core.data.crud.memo import crud_memo
 from app.core.data.crud.user import crud_user
@@ -9,73 +8,12 @@ from app.core.data.dto.annotation_document import AnnotationDocumentRead
 from app.core.data.dto.code import CodeRead
 from app.core.data.dto.memo import MemoRead
 from app.core.data.dto.project import ProjectRead
-from app.core.data.dto.user import (
-    UserAuthorizationHeaderData,
-    UserCreate,
-    UserLogin,
-    UserRead,
-    UserUpdate,
-)
-from app.core.mail.mail_service import MailService
-from app.core.security import generate_jwt
-from fastapi import APIRouter, Depends, HTTPException
-from fastapi.security import OAuth2PasswordRequestForm
+from app.core.data.dto.user import UserRead, UserUpdate
+from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 
-# no authentication dependency for the whole group here,
-# as login and register need to be accessible withaut a token
-router = APIRouter(prefix="/user")
+router = APIRouter(prefix="/user", dependencies=[Depends(get_current_user)])
 tags = ["user"]
-
-
-@router.put(
-    "",
-    tags=tags,
-    response_model=UserRead,
-    summary="Registers a new User",
-    description="Registers a new User and returns it with the generated ID.",
-)
-async def register(
-    *, db: Session = Depends(get_db_session), user: UserCreate
-) -> Optional[UserRead]:
-    db_user = crud_user.read_by_email(db, email=user.email)
-    if db_user:
-        raise HTTPException(
-            status_code=400,
-            detail="The user with this username already exists in the system.",
-        )
-
-    db_user = crud_user.create(db=db, create_dto=user)
-    await MailService().send_welcome_mail(user=UserRead.from_orm(db_user))
-    return UserRead.from_orm(db_user)
-
-
-@router.post(
-    "/login",
-    tags=tags,
-    response_model=UserAuthorizationHeaderData,
-    summary="Returns the JWT access token for the provided user login data",
-    description=(
-        "Returns the JWT access token for the provided user login data if the login was successful. "
-        "This is usually only called from an OAuth2 client!"
-    ),
-)
-async def login(
-    *,
-    db: Session = Depends(get_db_session),
-    user_login_form_data: OAuth2PasswordRequestForm = Depends()
-) -> UserAuthorizationHeaderData:
-    user_login = UserLogin(
-        username=user_login_form_data.username,
-        password=user_login_form_data.password,
-    )
-    user = crud_user.authenticate(db=db, user_login=user_login)
-    if not user:
-        raise credentials_exception
-
-    return UserAuthorizationHeaderData(
-        access_token=generate_jwt(user), token_type="bearer"
-    )
 
 
 @router.get(
@@ -95,7 +33,6 @@ async def get_me(*, user: UserRead = Depends(get_current_user)) -> Optional[User
     response_model=Optional[UserRead],
     summary="Returns the User",
     description="Returns the User with the given ID if it exists",
-    dependencies=[Depends(get_current_user)],
 )
 async def get_by_id(
     *, db: Session = Depends(get_db_session), user_id: int
@@ -110,7 +47,6 @@ async def get_by_id(
     response_model=List[UserRead],
     summary="Returns all Users",
     description="Returns all Users that exist in the system",
-    dependencies=[Depends(get_current_user)],
 )
 async def get_all(
     *,
@@ -127,7 +63,6 @@ async def get_all(
     response_model=Optional[UserRead],
     summary="Updates the User",
     description="Updates the User with the given ID if it exists",
-    dependencies=[Depends(get_current_user)],
 )
 async def update_by_id(
     *, db: Session = Depends(get_db_session), user_id: int, user: UserUpdate
@@ -142,7 +77,6 @@ async def update_by_id(
     response_model=Optional[UserRead],
     summary="Removes the User",
     description="Removes the User with the given ID if it exists",
-    dependencies=[Depends(get_current_user)],
 )
 async def delete_by_id(
     *, db: Session = Depends(get_db_session), user_id: int
@@ -157,7 +91,6 @@ async def delete_by_id(
     response_model=List[ProjectRead],
     summary="Returns all Projects of the User",
     description="Returns all Projects of the User with the given ID",
-    dependencies=[Depends(get_current_user)],
 )
 async def get_user_projects(
     *, user_id: int, db: Session = Depends(get_db_session)
@@ -173,7 +106,6 @@ async def get_user_projects(
     response_model=List[CodeRead],
     summary="Returns all Codes of the User",
     description="Returns all Codes of the User with the given ID",
-    dependencies=[Depends(get_current_user)],
 )
 async def get_user_codes(
     *, user_id: int, db: Session = Depends(get_db_session)
@@ -189,7 +121,6 @@ async def get_user_codes(
     response_model=List[MemoRead],
     summary="Returns all Memos of the User",
     description="Returns all Memos of the User with the given ID",
-    dependencies=[Depends(get_current_user)],
 )
 async def get_user_memos(
     *, user_id: int, db: Session = Depends(get_db_session)
@@ -208,7 +139,6 @@ async def get_user_memos(
     response_model=List[AnnotationDocumentRead],
     summary="Returns all Adocs of the User",
     description="Returns all Adocs of the User with the given ID",
-    dependencies=[Depends(get_current_user)],
 )
 async def get_user_adocs(
     *, user_id: int, db: Session = Depends(get_db_session)
@@ -226,7 +156,6 @@ async def get_user_adocs(
     response_model=List[AnnotationDocumentRead],
     summary="Returns sdoc ids of sdocs the User recently modified (annotated)",
     description="Returns the top k sdoc ids that the User recently modified (annotated)",
-    dependencies=[Depends(get_current_user)],
 )
 async def recent_activity(
     *, user_id: int, k: int, db: Session = Depends(get_db_session)
