+++ backend/src/configs/default.yaml
@@ -9,7 +9,7 @@ api:
   max_upload_file_size: 104857600 # 100 MB in bytes
   auth:
     jwt:
-      token_url: /user/login
+      token_url: /authentication/login
       ttl: ${oc.env:JWT_TTL, 600} # seconds
       algo: ${oc.env:JWT_ALGO, HS256}
       secret: ${oc.env:JWT_SECRET, "c7c802ec8df3f2661a3c64d7f617be89e93cb334fb282409e1df555543742bf2"}
