+++ frontend/src/api/openapi/index.ts
@@ -24,8 +24,8 @@ export type { BBoxAnnotationRead } from "./models/BBoxAnnotationRead";
 export type { BBoxAnnotationReadResolvedCode } from "./models/BBoxAnnotationReadResolvedCode";
 export type { BBoxAnnotationUpdateWithCodeId } from "./models/BBoxAnnotationUpdateWithCodeId";
 export type { Body_analysis_code_frequencies } from "./models/Body_analysis_code_frequencies";
+export type { Body_authentication_login } from "./models/Body_authentication_login";
 export type { Body_project_upload_project_sdoc } from "./models/Body_project_upload_project_sdoc";
-export type { Body_user_login } from "./models/Body_user_login";
 export type { CodeCreate } from "./models/CodeCreate";
 export type { CodeFrequency } from "./models/CodeFrequency";
 export type { CodeOccurrence } from "./models/CodeOccurrence";
@@ -114,6 +114,7 @@ export type { WhiteboardUpdate } from "./models/WhiteboardUpdate";
 export { AnalysisService } from "./services/AnalysisService";
 export { AnalysisTableService } from "./services/AnalysisTableService";
 export { AnnotationDocumentService } from "./services/AnnotationDocumentService";
+export { AuthenticationService } from "./services/AuthenticationService";
 export { BboxAnnotationService } from "./services/BboxAnnotationService";
 export { CodeService } from "./services/CodeService";
 export { CrawlerService } from "./services/CrawlerService";
