+++ frontend/src/api/openapi/services/UserService.ts
@@ -2,12 +2,9 @@
 /* tslint:disable */
 /* eslint-disable */
 import type { AnnotationDocumentRead } from "../models/AnnotationDocumentRead";
-import type { Body_user_login } from "../models/Body_user_login";
 import type { CodeRead } from "../models/CodeRead";
 import type { MemoRead } from "../models/MemoRead";
 import type { ProjectRead } from "../models/ProjectRead";
-import type { UserAuthorizationHeaderData } from "../models/UserAuthorizationHeaderData";
-import type { UserCreate } from "../models/UserCreate";
 import type { UserRead } from "../models/UserRead";
 import type { UserUpdate } from "../models/UserUpdate";
 
@@ -16,74 +13,6 @@ import { OpenAPI } from "../core/OpenAPI";
 import { request as __request } from "../core/request";
 
 export class UserService {
-  /**
-   * Returns all Users
-   * Returns all Users that exist in the system
-   * @returns UserRead Successful Response
-   * @throws ApiError
-   */
-  public static getAll({
-    skip,
-    limit,
-  }: {
-    /**
-     * The number of elements to skip (offset)
-     */
-    skip?: number;
-    /**
-     * The maximum number of returned elements
-     */
-    limit?: number;
-  }): CancelablePromise<Array<UserRead>> {
-    return __request(OpenAPI, {
-      method: "GET",
-      url: "/user",
-      query: {
-        skip: skip,
-        limit: limit,
-      },
-      errors: {
-        422: `Validation Error`,
-      },
-    });
-  }
-
-  /**
-   * Registers a new User
-   * Registers a new User and returns it with the generated ID.
-   * @returns UserRead Successful Response
-   * @throws ApiError
-   */
-  public static register({ requestBody }: { requestBody: UserCreate }): CancelablePromise<UserRead> {
-    return __request(OpenAPI, {
-      method: "PUT",
-      url: "/user",
-      body: requestBody,
-      mediaType: "application/json",
-      errors: {
-        422: `Validation Error`,
-      },
-    });
-  }
-
-  /**
-   * Returns the JWT access token for the provided user login data
-   * Returns the JWT access token for the provided user login data if the login was successful. This is usually only called from an OAuth2 client!
-   * @returns UserAuthorizationHeaderData Successful Response
-   * @throws ApiError
-   */
-  public static login({ formData }: { formData: Body_user_login }): CancelablePromise<UserAuthorizationHeaderData> {
-    return __request(OpenAPI, {
-      method: "POST",
-      url: "/user/login",
-      formData: formData,
-      mediaType: "application/x-www-form-urlencoded",
-      errors: {
-        422: `Validation Error`,
-      },
-    });
-  }
-
   /**
    * Returns the current user
    * Returns the current (logged in) user
@@ -156,6 +85,38 @@ export class UserService {
       },
       body: requestBody,
       mediaType: "application/json",
+      errors: {
+        422: `Validation Error`,
+      },
+    });
+  }
+
+  /**
+   * Returns all Users
+   * Returns all Users that exist in the system
+   * @returns UserRead Successful Response
+   * @throws ApiError
+   */
+  public static getAll({
+    skip,
+    limit,
+  }: {
+    /**
+     * The number of elements to skip (offset)
+     */
+    skip?: number;
+    /**
+     * The maximum number of returned elements
+     */
+    limit?: number;
+  }): CancelablePromise<Array<UserRead>> {
+    return __request(OpenAPI, {
+      method: "GET",
+      url: "/user",
+      query: {
+        skip: skip,
+        limit: limit,
+      },
       errors: {
         422: `Validation Error`,
       },
